refactor: move math utility functions to own module
