+++ module/data/item/weapon.mjs
@@ -1,6 +1,6 @@
 import PhysicalItemTemplate from "./templates/physical-item.mjs";
 import ItemDescriptionTemplate from "./templates/item-description.mjs";
-import { filterObject, inRange, sum } from "../../utils.mjs";
+import { filterObject } from "../../utils.mjs";
 import DamageRollOptions from "../roll/damage.mjs";
 import RollableTemplate from "./templates/rollable.mjs";
 import CombatDamageWorkflow from "../../workflows/workflow/damage-workflow.mjs";
@@ -8,6 +8,7 @@ import { SYSTEM_TYPES } from "../../constants/constants.mjs";
 import * as ACTORS from "../../config/actors.mjs";
 import * as COMBAT from "../../config/combat.mjs";
 import * as ITEMS from "../../config/items.mjs";
+import { inRange, sum } from "../../utils/math.mjs";
 
 
 /**
